feat: made cluster server working, removed listener to separate file, created new user DB
