--- src/server/database.json
@@ -1,6 +1,0 @@
-{
-    "users": [
-      { "id": "a502c7ec-5592-4d24-b6e9-ad20638b85f6", "username": "Student1", "age": 20, "hobbies": ["football", "chess"] },
-      { "id": "b015cf2b-a6f3-4aa3-8185-83670d1f3fa1", "username": "Reviewer1", "age": 22, "hobbies": ["basketball", "piano"] }
-    ]
-  }
